# pos-backend/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: POS Papelería — API REST
  version: 1.0.0
  description: |
    API del sistema de punto de venta para papelería.
    Desarrollada con Node.js/Express + Supabase (PostgreSQL).

    ## Autenticación
    Todos los endpoints (salvo `/auth/login`) requieren un **JWT Bearer token**
    obtenido desde `POST /api/auth/login`.

    ```
    Authorization: Bearer <token>
    ```

    ## Roles
    | Rol | Permisos |
    |---|---|
    | `cajero` | ventas, consulta de productos e inventario |
    | `administrador` | todo lo anterior + productos, inventario, corte de caja |

    ## Formato de respuesta exitosa
    ```json
    { "success": true, "data": { ... } }
    ```

    ## Formato de error
    ```json
    { "success": false, "error": "Mensaje descriptivo" }
    ```
servers:
  - url: http://localhost:3000/api
    description: Desarrollo local
  - url: http://192.168.1.100:3000/api
    description: Red local ITH

# ──────────────────────────────────────────────────────────────────────────────
# Security scheme
# ──────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido desde POST /auth/login. Expira en 10 horas.

  # ── Schemas reutilizables ────────────────────────────────────────────────────
  schemas:
...
    Error400:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Mensaje de validación" }
    Error401:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Token inválido o expirado" }
    Error403:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Se requiere rol administrador" }
    Error404:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Recurso no encontrado" }
...
    ItemVenta:
      type: object
      required: [id_producto, cantidad]
      properties:
        id_producto: { type: integer, minimum: 1, example: 7 }
        cantidad:    { type: integer, minimum: 1, example: 2 }
    DetalleVenta:
      type: object
      properties:
        id_detalle:      { type: integer, example: 55 }
        id_venta:        { type: integer, example: 12 }
        id_producto:     { type: integer, example: 7 }
        cantidad:        { type: integer, example: 2 }
        precio_unitario: { type: number,  format: decimal, example: 35.00 }
        descuento_linea: { type: number,  format: decimal, example: 0 }
        subtotal_linea:  { type: number,  format: decimal, example: 70.00 }
        productos:
          type: object
          properties:
            nombre:        { type: string, example: "Cuaderno universitario 100 hojas" }
            codigo_barras: { type: string, example: "7501000000007" }
    Venta:
      type: object
      properties:
        id_venta:           { type: integer, example: 12 }
        folio:              { type: string,  example: "VTA-20260624-0012" }
        fecha_hora:         { type: string,  format: date-time }
        id_usuario:         { type: integer, example: 3 }
        subtotal:           { type: number,  format: decimal, example: 95.50 }
        descuento:          { type: number,  format: decimal, example: 0 }
        total:              { type: number,  format: decimal, example: 95.50 }
        metodo_pago:
          type: string
          enum: [efectivo, tarjeta, mixto]
          example: efectivo
        monto_recibido:     { type: number,  format: decimal, example: 100.00 }
        cambio:             { type: number,  format: decimal, example: 4.50 }
        estado:
          type: string
          enum: [completada, cancelada, devolucion]
          example: completada
        motivo_cancelacion: { type: string,  nullable: true }
        id_corte:           { type: integer, nullable: true }
        usuarios:
          type: object
          properties:
            nombre: { type: string, example: "María López" }
        detalle_venta:
          type: array
          items:
            $ref: '#/components/schemas/DetalleVenta'
...
security:
  - bearerAuth: []

# ──────────────────────────────────────────────────────────────────────────────
# Tags
# ──────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /ventas:
    post:
      tags: [Ventas]
      summary: Registrar venta
      description: |
        Crea una venta completa: cabecera, detalle y descuento de stock.

        **Flujo interno:**
        1. Verifica stock disponible de cada producto
        2. Valida autorizador si `descuento > 20%` del subtotal
        3. Inserta cabecera en `ventas`
        4. Inserta líneas en `detalle_venta`
        5. Llama a RPC `decrementar_stock` por cada producto
        6. Registra movimientos en `movimientos_inventario`

        **Errores de negocio:**
        - Stock insuficiente → 400 con el nombre del producto y el disponible exacto
        - Descuento >20% sin autorizador → 400
        - Autorizador sin rol administrador → 400
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/ItemVenta' }
                metodo_pago:
                  type: string
                  enum: [efectivo, tarjeta, mixto]
                  default: efectivo
                monto_recibido:
                  type: number
                  minimum: 0
                  example: 100.00
                  description: Requerido cuando metodo_pago = efectivo o mixto
                descuento:
                  type: number
                  minimum: 0
                  default: 0
                  example: 0
                  description: Monto fijo de descuento en pesos
                id_autorizador:
                  type: integer
                  example: 1
                  description: Obligatorio cuando el descuento supera el 20% del subtotal
            examples:
              venta_efectivo:
                summary: Venta en efectivo con dos productos
                value:
                  items:
                    - { id_producto: 7,  cantidad: 2 }
                    - { id_producto: 12, cantidad: 3 }
                  metodo_pago: efectivo
                  monto_recibido: 200.00
                  descuento: 0
              venta_tarjeta:
                summary: Venta en tarjeta
                value:
                  items:
                    - { id_producto: 7, cantidad: 1 }
                  metodo_pago: tarjeta
              venta_con_descuento:
                summary: Venta con descuento del 15% (sin autorizador)
                value:
                  items:
                    - { id_producto: 7, cantidad: 4 }
                  metodo_pago: efectivo
                  monto_recibido: 120.00
                  descuento: 21.00
              venta_descuento_mayor:
                summary: Venta con descuento del 30% (requiere autorizador)
                value:
                  items:
                    - { id_producto: 7, cantidad: 4 }
                  metodo_pago: efectivo
                  monto_recibido: 100.00
                  descuento: 42.00
                  id_autorizador: 1
      responses:
        '201':
          description: Venta registrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: '#/components/schemas/Venta' }
              example:
                success: true
                data:
                  id_venta: 12
                  folio: "VTA-20260624-0012"
                  subtotal: 95.50
                  descuento: 0
                  total: 95.50
                  metodo_pago: efectivo
                  monto_recibido: 100.00
                  cambio: 4.50
                  estado: completada
        '400':
          description: Error de negocio
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error400' }
              examples:
                stock_cero:
                  summary: Producto sin stock
                  value: { success: false, error: 'Stock insuficiente para "Bolígrafo azul": disponible 0' }
                stock_insuficiente:
                  summary: Cantidad pedida supera el stock
                  value: { success: false, error: 'Stock insuficiente para "Lápiz HB": disponible 2' }
                descuento_sin_autorizar:
                  summary: Descuento alto sin autorizador
                  value: { success: false, error: "Descuentos mayores al 20% requieren id_autorizador con rol administrador" }
                autorizador_invalido:
                  summary: Autorizador sin rol administrador
                  value: { success: false, error: "El autorizador debe tener rol administrador" }
        '401': { description: No autenticado, content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }

    get:
      tags: [Ventas]
      summary: Listar ventas
      description: Devuelve ventas con filtros opcionales. Ordenadas por fecha descendente.
      parameters:
        - name: fecha
          in: query
          description: "Filtrar por fecha exacta (formato: YYYY-MM-DD)"
          schema: { type: string, format: date, example: "2026-06-24" }
        - name: estado
          in: query
          description: Filtrar por estado
          schema:
            type: string
            enum: [completada, cancelada, devolucion]
        - name: id_corte
          in: query
          description: Filtrar por ID de corte de caja
          schema: { type: integer, example: 5 }
      responses:
        '200':
          description: Lista de ventas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Venta' }
        '401': { description: No autenticado, content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
  /ventas/{id}:
    get:
      tags: [Ventas]
      summary: Obtener venta con detalle completo
      description: Devuelve la cabecera de la venta, los datos del cajero y todas las líneas del detalle con nombre y código de barras del producto.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 12 }
      responses:
        '200':
          description: Venta encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: '#/components/schemas/Venta' }
        '404':
          description: Venta no encontrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error404' }
              example: { success: false, error: "Venta no encontrada" }
        '401': { description: No autenticado, content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
  /ventas/{id}/ticket:
    get:
      tags: [Ventas]
      summary: Descargar ticket PDF
      description: Genera y descarga el ticket de venta en formato PDF.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 12 }
      responses:
        '200':
          description: Archivo PDF del ticket
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="ticket-VTA-20260624-0012.pdf"'
        '404': { description: Venta no encontrada, content: { application/json: { schema: { $ref: '#/components/schemas/Error404' } } } }
        '401': { description: No autenticado,      content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
  /ventas/{id}/cancelar:
    post:
      tags: [Ventas]
      summary: Cancelar venta
      description: |
        Cancela una venta completada y **reintegra el stock** de cada producto.

        **Requiere rol administrador.**

        Registra un movimiento de tipo `ajuste` en `movimientos_inventario`
        por cada producto de la venta.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 12 }
      responses:
        '200':
          description: Venta cancelada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: '#/components/schemas/Venta' }
              example:
                success: true
                data:
                  id_venta: 12
                  folio: "VTA-20260624-0012"
                  estado: cancelada
                  motivo_cancelacion: "Cancelada por administrador"
        '400':
          description: La venta ya está cancelada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error400' }
              example: { success: false, error: "La venta ya está cancelada" }
        '401': { description: No autenticado,   content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
        '403': { description: Rol insuficiente, content: { application/json: { schema: { $ref: '#/components/schemas/Error403' } } } }
        '404': { description: Venta no encontrada, content: { application/json: { schema: { $ref: '#/components/schemas/Error404' } } } }
  /ventas/{id}/devolucion:
    post:
      tags: [Ventas]
      summary: Procesar devolución parcial
      description: |
        Registra la devolución de uno o más productos de una venta.

        - Solo se devuelven los artículos incluidos en `items`.
        - El stock de cada artículo devuelto se incrementa vía RPC `incrementar_stock`.
        - Se registra un movimiento tipo `devolucion` por cada producto.
        - El estado de la venta cambia a `devolucion`.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 12 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/ItemVenta' }
            examples:
              devolucion_parcial:
                summary: Devolver 1 de 2 cuadernos comprados
                value:
                  items:
                    - { id_producto: 7, cantidad: 1 }
      responses:
        '201':
          description: Devolución procesada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: '#/components/schemas/Venta' }
              example:
                success: true
                data:
                  id_venta: 12
                  folio: "VTA-20260624-0012"
                  estado: devolucion
        '400':
          description: Error de negocio en devolución
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error400' }
              examples:
                cantidad_excedida:
                  value: { success: false, error: "Cantidad a devolver (5) supera la vendida (2)" }
                producto_ajeno:
                  value: { success: false, error: "Producto 99 no pertenece a esta venta" }
                venta_cancelada:
                  value: { success: false, error: "No se puede devolver una venta cancelada" }
        '401': { description: No autenticado,     content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
        '404': { description: Venta no encontrada, content: { application/json: { schema: { $ref: '#/components/schemas/Error404' } } } }

  # ═══════════════════════════════════════════════════════════════════════════
  # INVENTARIO
  # ═══════════════════════════════════════════════════════════════════════════
...